feature/saveitem accessing authenticated user added

+++ backend/otmz-api/src/main/java/com/patandmat/otmz/domain/imageFile/application/ImageFileServiceImpl.java
@@ -8,10 +8,12 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.awt.*;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -54,7 +56,16 @@ public class ImageFileServiceImpl implements ImageFileService {
     }
 
     @Override
-    public void delete(String path) throws IOException {
+    @Transactional
+    public void delete(Long id) throws IOException {
+        Optional<ImageFile> imageFileOptional = imageFileRepository.findById(id);
+        if (!imageFileOptional.isPresent()) return;
+
+        ImageFile imageFile = imageFileOptional.get();
+        String path = imageFile.getPath();
+
+        imageFileRepository.delete(imageFile);
+
         File f = new File(absolutePath+path);
         if (!f.exists()) return;
         Files.delete(Path.of(absolutePath,path));
@@ -82,7 +93,6 @@ public class ImageFileServiceImpl implements ImageFileService {
             fileArray = outputStream.toByteArray();
             inputStream.close();
             outputStream.close();
-
         }
         catch (IOException e) {
             throw new Exception("파일을 변환하는데 문제가 발생했습니다.");
